# offer-demo-service/src/test/resources/yaml/swagger-v1.yaml
swagger: '2.0'
info:
  description: RESTful API for Offers
  version: '1.0'
  title: Offers REST API
  contact:
    name: Frederic Thomas
    url: 'https://github.com/doublefx'
    email: fthomas@apache.org
  license:
    name: Apache License Version 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
tags:
  - name: offer-controller
    description: Offer Controller
paths:
  /offer/create:
    post:
      tags:
        - offer-controller
      summary: Create one offer.
      operationId: create
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: offer
          description: offer
          required: true
          schema:
            $ref: '#/definitions/Offer'
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /offer/disable:
    get:
      tags:
        - offer-controller
      summary: Disable an offer.
      operationId: disable
      produces:
        - application/json
      parameters:
        - name: offerId
          in: query
          description: offerId
          required: true
          type: integer
          format: int64
          allowEmptyValue: false
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /offer/getAll:
    get:
      tags:
        - offer-controller
      summary: Get all offers not expired and not disabled.
      operationId: getAll
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Offer'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
definitions:
  Offer:
    type: object
    properties:
      description:
        type: string
        example: This is the best offer
      expiryDate:
        type: string
        format: date-time
        example: "2039-06-16T06:44:17Z"
      id:
        type: integer
        format: int64
        example: 1
      price:
        type: number
        example: 10.50
    title: Offer